+++ UCR/App.config
@@ -1,8 +1,8 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <configuration>
   <configSections>
-    <sectionGroup name="userSettings" type="System.Configuration.UserSettingsGroup, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" >
-      <section name="HidWizards.UCR.Properties.Settings" type="System.Configuration.ClientSettingsSection, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" allowExeDefinition="MachineToLocalUser" requirePermission="false" />
+    <sectionGroup name="applicationSettings" type="System.Configuration.ApplicationSettingsGroup, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" >
+      <section name="HidWizards.UCR.Properties.Settings" type="System.Configuration.ClientSettingsSection, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" requirePermission="false" />
     </sectionGroup>
   </configSections>
   <startup>
@@ -13,7 +13,7 @@
       <probing privatePath="lib;libs" />
     </assemblyBinding>
   </runtime>
-  <userSettings>
+  <applicationSettings>
     <HidWizards.UCR.Properties.Settings>
       <setting name="StartMinimized" serializeAs="String">
         <value>False</value>
@@ -25,5 +25,5 @@
         <value>250, 400</value>
       </setting>
     </HidWizards.UCR.Properties.Settings>
-  </userSettings>
+  </applicationSettings>
 </configuration>
